Finishing upvote
I thought that adding a new element in the list was not a satisfying
solution. Of course, it makes it more visible, but the entries get a few
noisy and you sometimes have to scroll a lot to find the upvote
suggestion. So now, when you upvote, the counter of vote simply
increment by one.
By the way, I changed the List of Suggestion to a Dict Suggestion.title
Suggestion. I hope that there will not be two suggestions with the same
title, or my stuff will be pretty fucked up.

Thank you for reading this commit message.

TL;DR: Finished upvote and List of Suggestion is now Dict of Suggestion